requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
opencv-python>=4.8.0
loguru>=0.7.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.25
Pillow>=10.0.0
requests
insightface==0.7.3
onnxruntime
fpdf2